Allow "arbitrary" arguments to corrplot
As of now, you can only supply matrices to corrplot; however, to my understanding, corrplot is applied relatively early in the plotting pipeline, and cannot be accessed through a seriestype.  That being said, it would be nice for user-defined types which are "`corrplot`-able" to be easily plottable through a function overload.

Therefore, I've introduced here a function `to_corrplot_matrix(mat)`, which can be overloaded by the user for their custom types.  It must return a matrix of a form which can be ingested by `corrplot`.

It would be great if there's a recipe-based solution to this that I'm missing; however, this kind of "multiple-dispatch" based recipe seems hard to implement without this kind of workaround (please tell me if I'm wrong here).  Perhaps it would be possible to introduce a plot-type-based type conversion recipe syntax, as well?