Added kw to set whiskers at quantile values
If the keyword `quantile_range = true`, the value supplied in range gives the lower quantile that will be inside the whiskers, the upper quantile being 1-range. The value of range is defaulted to 0.05 if `quantile_range` is true, otherwise the usual 1.5 is kept. If the range is higher than 1 and `quantile_value` is true, the range is defaulted to 1, as if range was equal to 0. 

I added this since it is common in plotting packages to have this functionality.