Fix dendrogram (#447)
* Fix dendrogram

Hi!

This fixes a minor bug I introduced in https://github.com/JuliaPlots/StatsPlots.jl/pull/356 to match the Wikipedia example. I'm sorry for that. In that example, the merged nodes' height is the branch length, which is half of the distance between the merged nodes for the ultrametric tree. That's is problematic in this context, where `heights[i]` is already the `height` at which the merge node `i' should be placed. Both `Clustering.jl` and `R` considered that the height is the actual distance rather than the brach length, which makes much sense in combination with `hline!` and `cutree` (how I've realised this bug). With this simple fix, our plot matches the R plot :) 

This also fixes https://github.com/JuliaPlots/StatsPlots.jl/issues/224 

Many thanks for considering my PR,

Best regards,

* Add test to avoid regression

`ys` values make sense with the distances in the wiki article example and the visual output for those coordinates is identical to *R*: `plot(hclust(as.dist(wiki_example), method="complete"), hang=-1)`